+++ demo/headings.ts
@@ -2,15 +2,8 @@ export { headings }
 export { headingsDetached }
 export { categories }
 
-import {
-  iconScroll,
-  iconCompass,
-  iconGear,
-  type HeadingDefinition,
-  type HeadingDetachedDefinition,
-  type Config,
-  iconLoudspeaker,
-} from '@brillout/docpress'
+import type { HeadingDefinition, HeadingDetachedDefinition, Config } from '@brillout/docpress'
+import { iconScroll, iconCompass, iconGear, iconLoudspeaker } from '@brillout/docpress' with { type: 'vike:pointer' }
 
 const categories: Config['categories'] = [
   //